commit html update

+++ todolist/models.py
@@ -1,6 +1,6 @@
 from django.db import models
 from django.db.models.fields import FloatField, IntegerField, DecimalField
-from djmoney.money import Money
+
 
 # Create your models here.
 class Task(models.Model):
@@ -17,7 +17,7 @@ class ValorIntroducido(models.Model):
         return str(self.precio)
 
 class Resultados(models.Model):
-    resultado = IntegerField(default=0)
+    resultado = FloatField(default=0)
     def __str__(self):
         return str(self.resultado)
